Move Shared pdf package to util
