+++ r2-streamer/src/main/java/org/readium/r2/streamer/parser/readium/LcpdfPositionsService.kt
@@ -12,7 +12,7 @@ package org.readium.r2.streamer.parser.readium
 import android.content.Context
 import org.readium.r2.shared.fetcher.Fetcher
 import org.readium.r2.shared.format.MediaType
-import org.readium.r2.shared.pdf.PdfDocument
+import org.readium.r2.shared.util.pdf.PdfDocument
 import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.Publication
